+++ SchoolBook.java
@@ -18,7 +18,7 @@ public class SchoolBook extends Book {
         pricePerDay = price;
     }
     @Override
-    public Price getPrice(int days) {
+    public Price getPrice(long days) {
         return Price.multiply(pricePerDay, days);
     }
 }
